Move RepoData from sphinx ext to anaconda
